sysusers: allow force reusing existing user/group IDs
On Debian/Ubuntu systems the default passwd/group files use a
slightly strange mapping. E.g. in passwd:
```
man:x:6:12::/var/cache/man:/sbin/nologin
```

and in group:
```
disk:x:6:
man:x:12:
```

This is not supported in systemd-sysusers right now because
sysusers will not re-use an existing uid/gid in its normal
mode of operation. Unfortunately this reuse is needed to
replicate the default Debian/Ubuntu users/groups.

This PR enforces reuse when the "uid:gid" syntax is used
to fix this.

I also added a test that replicates the Debian base-passwd
passwd/group file to ensure things are ok. The only issue
left is that on Debian "nologin" is located in /usr/sbin/nologin
but that is something I will discuss with the Debian people
first.

If there are compatibility concerns I'm happy to do it in a
different way, e.g. via a "force" prefix/postfix like
`u user uid:gid! -` or similar.