Make GetManagedObjects() return in descent order
Before this patch, calling GetManagedObjects() returned the objects in an
undefined and apparently random order. This means that randomly /foo/bar
would be returned before /foo. While this is technically not a violation of
the specification, it does appear to be different than all the
implementations that I tried which returned the objects predictably in
descent order. That is, /foo always appears before /foo/bar.

The existing code collects all child paths in a Set and then serializes
each one. This is what produces the undefined order. Instead, we modify the
code to simply recursively descend the tree and serialize each node as we
encounter it. This is not only the most logical behavior but it is also
more performant.

Fixes: #8008