systemctl: add a way to show X-* properties
Extra properties (and sections) are ignored by systemd, so can't be queried
with systemctl. Programs that use extra properties can use "systemctl cat" to
get access to all relevant files for an unit, but still need to handle some
configuration parsing issues (continuation lines, sections).

The new command provides a way to query the effective value of one or more
extra section/key pairs in one or more units. It does the section parsing, and
outputs only the last set values of each property (or all the values set in
case --full is passed). This reduces the requirements on external programs to
use X-* properties.