logind: effectively finalize/free a session after it was stopped
It was noticed that a session is not effectively freed, causing
a relevant memory consumption of logind. The simple test to
reproduce this issue is to run a ssh connection loop like this:

  while true; do ssh target_machine "whoami"; done

This causes continuous growth of memory consumption by logind
until it reaches the maximum allowed sessions threshold
(default to 8192).

This patch fixes it by effectively removing/freeing a session
in garbage collector, by checking if a session was already
stopped on manager_gc(), and in positive case, calling
session_finalize() and session_free().

Signed-off-by: Guilherme G. Piccoli <gpiccoli@canonical.com>