slice, scope: IgnoreOnIsolate=yes is already the default
`IgnoreOnIsolate=yes` is the default for slices and scopes.  So it's not
essential to set it on root.slice or init.scope.

We don't need to worry about a bad unit file configuration.  Any attempt
to stop these unit should fail, since we mark them as `perpetual`.

Also since init.scope cannot be stopped, there is no point setting
`KillSignal=SIGRTMIN+14`.  According to both documentation and testing,
KillSignal= does not affect the behaviour of `systemctl kill`.