meson: support the case glibc internally uses libidn2
On Fedora rawhide, we get the following warning:
```
../src/basic/socket-util.c: In function 'socknameinfo_pretty':
../src/basic/socket-util.c:742:13: warning: NI_IDN_USE_STD3_ASCII_RULES is deprecated
         r = getnameinfo(&sa->sa, salen, host, sizeof(host), NULL, 0, IDN_FLAGS);
             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
This is caused by the glibc patch that makes glibc use libidn2 internally.
The patch deprecates several defined values including `NI_IDN_USE_STD3_ASCII_RULES`.
The patch has not yet merged in the glibc upstream. But Fedora rawhide
has already enabled the patch.

This updates meson.build to detect the patch is applied or not,
and suppress the above warning message.