hwdb: Add new snd_hda_intel power-saving blacklist hwdb file
OpenSUSE has already been shipping kernels with
CONFIG_SND_HDA_POWER_SAVE_DEFAULT=1 for a while now and Fedora is going
to do the same starting with Fedora 28. In general this is a good thing
as it saves aprox. 0.4W of power when the audio codec is not active.

But on some devices with snd_hda_intel audio, this leads to a plopping or
crackling sound each time the codec is turned on/off.

This usually points to a problem in the hda-codec driver for the specific
external codec used together with the Intel HDA audio integrated into the
chipset. Unfortunately there are many different external codecs and even
more ways these could be wired up and then they can be paired with an
external amplifier which may have multiple enable pins, etc. So fixing
this is a tedious and time-consuming process which needs to be done per
board.

TL;DR: the undesired sounds when powersaving kicks in are typically a
kernel-bug but getting these fixed in a timely manner on all affected
models is not feasible.

Since CONFIG_SND_HDA_POWER_SAVE_DEFAULT=1 works problem free on aprox
95-99% of all models out there and since it gives a nice increase in
battery life we really want to have it on by default

So this commit implements plan B for devices on which this causes problems,
a blacklist in hwdb which turns of the power-saving on these specific
devices.

This initial commit adds only one model laptop to the hwdb, but I expect
more models to follow in the future.

Related bugzillas:
https://bugzilla.kernel.org/show_bug.cgi?id=198611
https://bugzilla.redhat.com/show_bug.cgi?id=1525561
https://bugzilla.redhat.com/show_bug.cgi?id=1525104
https://bugzilla.redhat.com/show_bug.cgi?id=1526678
https://bugzilla.redhat.com/show_bug.cgi?id=1533858