fstab-generator: don't warn on "noauto" symlinks
The fstab-generator tries to canonicalize symlinks. It's strange when
unsuccessful canonicalization is reported as error if the fstab entry
is explicitly marked as "noauto". In this case users assume that
systemd will not try to be smart on the fstab entry. Note that user
cannot do anything more than use "noauto"...

The patch does not change the current behavior, but suppress the error
message for "noauto" only

Addresses: https://bugzilla.redhat.com/show_bug.cgi?id=1537969