timedated: add possibility to set a different path for files
/etc is currently the path for localtime and adjtime symlink to
remove/create when updating the timezone/RTC but this isn't compatible
with read-only rootfs.

The proposed change is to make path to localtime/adjtime to be set at
build time so that we can select where timedated will look at (in
read-write filesystems) when doing operations on RTC/timezones.

We'll need to make sure the following links are set though since other
tools expect to find a symbolic link at /etc/localtime and /etc/adjtime:

/etc/localtime -> /given/path/localtime -> /usr/share/zoneinfo/...
/etc/adjtime -> /given/path/adjtime